+++ examples/sleep.rs
@@ -1,103 +1,113 @@
-//! This example is public domain.
+#![deny(unsafe_code)]
 
-use smelling_salts::{Device, Watcher};
-use std::future::Future;
-use std::mem::MaybeUninit;
-use std::os::raw;
-use std::pin::Pin;
-use std::task::{Context, Poll};
-use std::time::Duration;
+/// Timer module
+mod timer {
+    #![allow(unsafe_code)]
 
-#[repr(C)]
-struct TimeSpec {
-    sec: isize,
-    nsec: raw::c_long,
-}
+    use flume::Sender;
+    use smelling_salts::linux::{Device, Driver, RawDevice, Watcher};
+    use std::convert::TryInto;
+    use std::future::Future;
+    use std::mem::{self, MaybeUninit};
+    use std::os::raw;
+    use std::pin::Pin;
+    use std::ptr;
+    use std::sync::Once;
+    use std::task::{Context, Poll};
+    use std::time::Duration;
 
-#[repr(C)]
-struct ITimerSpec {
-    interval: TimeSpec,
-    value: TimeSpec,
-}
+    fn driver() -> &'static Driver {
+        static mut DRIVER: MaybeUninit<Driver> = MaybeUninit::uninit();
+        static ONCE: Once = Once::new();
+        unsafe {
+            ONCE.call_once(|| DRIVER = MaybeUninit::new(Driver::new()));
+            &*DRIVER.as_ptr()
+        }
+    }
 
-extern "C" {
-    fn timerfd_create(clockid: raw::c_int, flags: raw::c_int) -> raw::c_int;
-    fn timerfd_settime(
-        fd: raw::c_int,
-        flags: raw::c_int,
-        new_value: *const ITimerSpec,
-        old_value: *mut ITimerSpec,
-    ) -> raw::c_int;
-    fn read(fd: raw::c_int, buf: *mut u64, count: usize) -> isize;
-    fn close(fd: raw::c_int) -> raw::c_int;
-    fn __errno_location() -> *mut raw::c_int;
-}
+    #[repr(C)]
+    struct TimeSpec {
+        sec: isize,
+        nsec: raw::c_long,
+    }
 
-struct Sleep(Device, u64);
+    #[repr(C)]
+    struct ITimerSpec {
+        interval: TimeSpec,
+        value: TimeSpec,
+    }
 
-impl Sleep {
-    fn new(dur: Duration) -> Self {
-        let sec = dur.as_secs() as _;
-        let nsec = dur.subsec_nanos() as _;
+    extern "C" {
+        fn timerfd_create(clockid: raw::c_int, flags: raw::c_int) -> RawDevice;
+        fn timerfd_settime(
+            fd: RawDevice,
+            flags: raw::c_int,
+            new_value: *const ITimerSpec,
+            old_value: *mut ITimerSpec,
+        ) -> raw::c_int;
+        fn read(fd: RawDevice, buf: *mut u64, count: usize) -> isize;
+        fn close(fd: RawDevice) -> raw::c_int;
+    }
 
-        let timerfd = unsafe {
-            timerfd_create(1 /*Monotonic*/, 2048 /*Nonblock*/)
-        };
-        let x = unsafe {
-            timerfd_settime(
-                timerfd,
-                0,
-                &ITimerSpec {
-                    interval: TimeSpec { sec, nsec },
-                    value: TimeSpec { sec, nsec },
-                },
-                std::ptr::null_mut(),
-            )
-        };
-        assert_eq!(0, x);
+    struct TimerDriver(Sender<usize>, RawDevice);
 
-        Sleep(Device::new(timerfd, Watcher::new().input()), 0)
+    impl TimerDriver {
+        unsafe fn callback(&mut self) -> Option<()> {
+            let mut x = MaybeUninit::<u64>::uninit();
+            let v = read(self.1, x.as_mut_ptr(), mem::size_of::<u64>());
+            if v == mem::size_of::<u64>().try_into().unwrap()
+                && self.0.send(x.assume_init().try_into().unwrap()).is_err()
+            {
+                driver().discard(self.1);
+                let _ret = close(self.1);
+                assert_eq!(0, _ret);
+                return None;
+            }
+            Some(())
+        }
+    }
+
+    /// A `Timer` device future.
+    pub struct Timer(Device<usize>);
+
+    impl Timer {
+        /// Create a new `Timer`.
+        pub fn new(dur: Duration) -> Self {
+            // Create Monotonic (1), Non-Blocking (2048) Timer
+            let fd = unsafe { timerfd_create(1, 2048) };
+            let sec = dur.as_secs() as _;
+            let nsec = dur.subsec_nanos() as _;
+            let its = ITimerSpec {
+                interval: TimeSpec { sec, nsec },
+                value: TimeSpec { sec, nsec },
+            };
+            let _ret = unsafe { timerfd_settime(fd, 0, &its, ptr::null_mut()) };
+            assert_eq!(0, _ret);
+            let constructor = |sender| TimerDriver(sender, fd);
+            let callback = TimerDriver::callback;
+            let watcher = Watcher::new().input();
+            Self(driver().device(constructor, fd, callback, watcher))
+        }
+    }
+
+    impl Future for Timer {
+        type Output = usize;
+        fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<usize> {
+            Pin::new(&mut self.get_mut().0).poll(cx)
+        }
     }
 }
 
-impl Future for Sleep {
-    type Output = ();
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<()> {
-        // Queue
-        if self.1 != 0 {
-            self.1 -= 1;
-            return Poll::Ready(());
-        }
-        // Early return if a different device woke the executor.
-        if self.0.pending() {
-            return self.0.sleep(cx);
-        }
-        //
-        let mut x = MaybeUninit::<u64>::uninit();
-        let v = unsafe {
-            read(self.0.raw(), x.as_mut_ptr(), std::mem::size_of::<u64>())
-        };
-        if v > 0 {
-            self.1 += unsafe { x.assume_init() };
-            self.poll(cx)
-        } else {
-            self.0.sleep(cx)
-        }
-    }
-}
-
-impl Drop for Sleep {
-    fn drop(&mut self) {
-        assert_eq!(0, unsafe { close(self.0.stop()) });
-    }
-}
+// Export the `Timer` future.
+use timer::Timer;
 
 fn main() {
     pasts::block_on(async {
-        for _ in 0..5 {
-            println!("Sleeping for 1 second…");
-            Sleep::new(Duration::new(1, 0)).await;
+        let mut timer = Timer::new(std::time::Duration::from_secs_f32(1.0));
+        println!("Sleeping for 1 second 5 times…");
+        for i in 0..5 {
+            (&mut timer).await;
+            println!("Slept {} time(s)…", i + 1);
         }
     });
 }