+++ examples/wasm_timer/static/cala.js
@@ -1,3 +1,0 @@
-export {
-    cala_test: function() { alert("Hello, world!"); }
-}